+++ PTC_xcode/PTC_xcode/route_prediction/Route.cpp
@@ -66,7 +66,7 @@ Route* Route::copy() {
 	return route;
 }
 
-int Route::getGoalHash() {
+long int Route::getGoalHash() {
 	return this->goal->getHash();
 }
